refactor(http): explicit imports in app/ + split error responses out of util
Address the review's two app/ findings:

- util.rs was a grab-bag (its own doc was a noun list): extract the error/
  diagnostic response constructors into error_responses.rs. util.rs now holds
  request-inspection helpers + the owner/admin guards.
- Replace the `use super::*` + sibling-glob imports (e.g. proxy.rs globbed all
  of backend to use one symbol) with explicit named imports, matching
  aci/upstream, aci/verifier, and aggregator/upstream_config. Each file now
  states its real dependencies in its header.

No behavior change; cargo test --all-targets: 220 passed, 0 failed.
